+++ t2wml/wikification/utility_functions.py
@@ -5,16 +5,19 @@ from pathlib import Path
 from t2wml.utils.utilities import VALID_PROPERTY_TYPES
 from SPARQLWrapper.SPARQLExceptions import QueryBadFormed
 from t2wml.utils import t2wml_exceptions as T2WMLExceptions
-from t2wml.wikification.wikidata_provider import DictionaryProvider
-from t2wml.wikification.preloaded_properties import preloaded_properties
+from t2wml.wikification.wikidata_provider import KGTKFileProvider
 from t2wml.settings import t2wml_settings
 
 
+def get_default_provider():
+    default_kgtk=Path(__file__).parent / "preloaded_properties.tsv"
+    wikidata_provider = KGTKFileProvider(default_kgtk)
+    return wikidata_provider
 
 def get_provider():
     wikidata_provider = t2wml_settings.wikidata_provider
     if wikidata_provider is None:
-        wikidata_provider = DictionaryProvider(preloaded_properties)
+        wikidata_provider = get_default_provider()
         t2wml_settings.wikidata_provider = wikidata_provider
     return wikidata_provider
 
@@ -51,6 +54,46 @@ def validate_id(node_id):
             raise T2WMLExceptions.InvalidEntityDefinition("Custom entity ID Qnum where num<1 billion is not allowed: "+str(node_id))
     except ValueError: #conversion to int failed, is not Pnum or Qnum
         pass
+
+
+def kgtk_to_dict(file_path):
+    #if label is P2010050001 (datamart tag), the key is tags and a list is generated
+    input_dict=defaultdict(dict)
+    with open(file_path, 'r', encoding="utf-8") as f:
+        reader = csv.DictReader(f, delimiter="\t")
+        for row_dict in reader:
+            node1 = row_dict["node1"]
+            label = row_dict["label"]
+            value = row_dict["node2"]
+            if label == "P2010050001":
+                if "tags" in input_dict[node1]:
+                    input_dict[node1]["tags"].append(value)
+                else:
+                    input_dict[node1]["tags"]=[value]
+            else:
+                input_dict[node1][label]=value
+    return dict(input_dict)
+
+def dict_to_kgtk(in_dict, out_path):
+    tsv_dict_columns=["id", "node1", "label", "node2"]
+    tsv_dict_arr=[]
+    for node1, node_dict in in_dict.items():
+        for label, value in node_dict.items():
+            if label=="tags":
+                for index, tag in enumerate(value):
+                    tsv_dict_arr.append(dict(node1=node1, label="P2010050001", node2=tag, id=f'{node1}-{label}-{index}'))
+            else:
+                tsv_dict_arr.append(dict(node1=node1, label=label, node2=value, id=f'{node1}-{label}'))
+    with open(out_path, 'w', encoding="utf-8") as f:
+        dw= csv.DictWriter(f, tsv_dict_columns,
+                            restval="", delimiter="\t", lineterminator="\n",
+                            escapechar='', quotechar='',
+                            dialect=csv.unix_dialect, quoting=csv.QUOTE_NONE)
+        dw.writeheader()
+        for line in tsv_dict_arr:
+            dw.writerow(line)
+
+
 
 def add_entities_from_file(file_path: str, validate_ids=True):
     """load wikidata entries from a file and add them to the current WikidataProvider as defined in settings.